--- packages/app-core/src/AppFocus/index.ts
@@ -1,24 +1,0 @@
-import { types } from '@jbrowse/mobx-state-tree'
-
-/**
- * #stateModel AppFocusMixin
- * #category session
- */
-export function AppFocusMixin() {
-  return types
-    .model({
-      /**
-       * #property
-       * used to keep track of which view is in focus
-       */
-      focusedViewId: types.maybe(types.string),
-    })
-    .actions(self => ({
-      /**
-       * #action
-       */
-      setFocusedViewId(viewId: string) {
-        self.focusedViewId = viewId
-      },
-    }))
-}
