--- packages/app-core/src/Assemblies/index.ts
@@ -1,2 +1,0 @@
-export * from './SessionAssembliesMixin'
-export * from './TemporaryAssembliesMixin'
