+++ packages/app-core/src/index.ts
@@ -1,10 +1,6 @@
 export * from './ui'
-export * from './HistoryManagement'
-export * from './AppFocus'
 export * from './JBrowseConfig'
 export * from './JBrowseModel'
-export * from './Assemblies'
-export * from './RootMenu'
 export * from './menus'
 export * from './menuItems'
 export * from './types'
