+++ packages/product-core/src/RootModel/index.ts
@@ -1,5 +1,4 @@
 export * from './BaseRootModel'
-export * from './InternetAccounts'
 export * from './FormatAbout'
 export * from './FormatDetails'
 export * from './HierarchicalConfig'
