--- packages/product-core/src/Session/Themes.ts
@@ -1,87 +1,0 @@
-import { readConfObject } from '@jbrowse/core/configuration'
-import { createJBrowseTheme, defaultThemes } from '@jbrowse/core/ui'
-import { localStorageGetItem, localStorageSetItem } from '@jbrowse/core/util'
-import { addDisposer, types } from '@jbrowse/mobx-state-tree'
-import { autorun } from 'mobx'
-
-import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
-import type PluginManager from '@jbrowse/core/PluginManager'
-import type { IAnyStateTreeNode, Instance } from '@jbrowse/mobx-state-tree'
-import type { ThemeOptions } from '@mui/material'
-
-type ThemeMap = Record<string, ThemeOptions>
-
-interface ThemeMixinContext {
-  jbrowse: AnyConfigurationModel
-}
-
-/**
- * #stateModel ThemeManagerSessionMixin
- */
-export function ThemeManagerSessionMixin(_pluginManager: PluginManager) {
-  return types
-    .model({})
-    .volatile(() => ({
-      sessionThemeName: localStorageGetItem('themeName') || 'default',
-    }))
-    .views(s => {
-      const self = s as typeof s & ThemeMixinContext
-      return {
-        /**
-         * #method
-         */
-        allThemes(): ThemeMap {
-          const extraThemes = readConfObject(self.jbrowse, 'extraThemes')
-          return { ...defaultThemes, ...extraThemes }
-        },
-        /**
-         * #getter
-         */
-        get themeName() {
-          const { sessionThemeName } = s
-          const all = this.allThemes()
-          return all[sessionThemeName] ? sessionThemeName : 'default'
-        },
-        /**
-         * #getter
-         */
-        get theme() {
-          const configTheme = readConfObject(self.jbrowse, 'theme')
-          const all = this.allThemes()
-          return createJBrowseTheme(configTheme, all, this.themeName)
-        },
-      }
-    })
-    .actions(self => ({
-      /**
-       * #action
-       */
-      setThemeName(name: string) {
-        self.sessionThemeName = name
-      },
-      afterAttach() {
-        addDisposer(
-          self,
-          autorun(
-            function themeNameAutorun() {
-              localStorageSetItem('themeName', self.themeName)
-            },
-            { name: 'ThemeName' },
-          ),
-        )
-      },
-    }))
-}
-
-/** Session mixin MST type for a session that supports theming */
-export type SessionWithThemesType = ReturnType<typeof ThemeManagerSessionMixin>
-
-/** Instance of a session that has theming support */
-export type SessionWithThemes = Instance<SessionWithThemesType>
-
-/** Type guard for SessionWithThemes */
-export function isSessionWithThemes(
-  session: IAnyStateTreeNode,
-): session is SessionWithThemes {
-  return 'theme' in session
-}
