--- packages/product-core/src/Session/Tracks.ts
@@ -1,76 +1,0 @@
-import { types } from '@jbrowse/mobx-state-tree'
-
-import { BaseSessionModel, isBaseSession } from './BaseSession'
-import { ReferenceManagementSessionMixin } from './ReferenceManagement'
-
-import type PluginManager from '@jbrowse/core/PluginManager'
-import type {
-  AnyConfiguration,
-  AnyConfigurationModel,
-} from '@jbrowse/core/configuration'
-import type { IAnyStateTreeNode, Instance } from '@jbrowse/mobx-state-tree'
-
-/**
- * #stateModel TracksManagerSessionMixin
- * composed of
- * - BaseSessionModel
- * - ReferenceManagementSessionMixin
- */
-export function TracksManagerSessionMixin(pluginManager: PluginManager) {
-  return types
-    .compose(
-      'TracksManagerSessionMixin',
-      BaseSessionModel(pluginManager),
-      ReferenceManagementSessionMixin(pluginManager),
-    )
-    .views(self => ({
-      /**
-       * #getter
-       */
-      get tracks(): AnyConfigurationModel[] {
-        return self.jbrowse.tracks
-      },
-    }))
-    .actions(self => ({
-      /**
-       * #action
-       */
-      addTrackConf(trackConf: AnyConfiguration) {
-        return self.jbrowse.addTrackConf(trackConf)
-      },
-
-      /**
-       * #action
-       */
-      deleteTrackConf(trackConf: AnyConfigurationModel) {
-        const callbacksToDereferenceTrack: (() => void)[] = []
-        const dereferenceTypeCount: Record<string, number> = {}
-        const referring = self.getReferring(trackConf)
-        self.removeReferring(
-          referring,
-          trackConf,
-          callbacksToDereferenceTrack,
-          dereferenceTypeCount,
-        )
-        for (const cb of callbacksToDereferenceTrack) {
-          cb()
-        }
-        if (self.adminMode) {
-          return self.jbrowse.deleteTrackConf(trackConf)
-        }
-      },
-    }))
-}
-
-/** Session mixin MST type for a session that has tracks */
-export type SessionWithTracksType = ReturnType<typeof TracksManagerSessionMixin>
-
-/** Instance of a session that has tracks */
-export type SessionWithTracks = Instance<SessionWithTracksType>
-
-/** Type guard for SessionWithTracks */
-export function isSessionWithTracks(
-  thing: IAnyStateTreeNode,
-): thing is SessionWithTracks {
-  return isBaseSession(thing) && 'tracks' in thing
-}
