+++ packages/product-core/src/Session/index.ts
@@ -1,10 +1,2 @@
-export * from './ReferenceManagement'
-export * from './Connections'
-export * from './DrawerWidgets'
-export * from './DialogQueue'
-export * from './Themes'
-export * from './Tracks'
-export * from './MultipleViews'
 export * from './BaseSession'
-export * from './SessionTracks'
-export * from './BaseSession'
+export * from './types'
